refactor(access): move beta-access, access-model and account-status into the module — and re-arm three guards the move disarmed (KAN-415 tail)
UI-Change-Approved: KAN-415

    src/lib/beta-access/{email,flow,invite-cookie,invite-link}.ts
                                    -> src/modules/access/beta-access/
    src/lib/access-model/{index,access-transitions}.ts
                                    -> src/modules/access/access-model/
    src/lib/account-status.ts       -> src/modules/access/account-status.ts

All seven are byte-identical renames (100%). 43 files rewritten across the @/
alias, prose, and the path-anchored artefacts below.

THREE GUARDS WENT DEAD THE MOMENT THE FILES MOVED

Measured before fixing anything, by running the guards against the staged move:

    .github/signup-surface.paths:37   src/lib/beta-access/**        -> matched nothing
    .github/CODEOWNERS:37            /src/lib/beta-access/          -> matched nothing
    check-ui-copy-ownership.sh:65     src/lib/beta-access/email.ts   -> matched nothing

The first is CTL-013, gotcha #32 — the gate that decides whether a promote must
run the signup E2E, and whose header says the daily soak deliberately never
exercises signup, so when it says CLEAN nothing else is looking.

The third is KAN-411 founder ownership. Measured directly:

    src/lib/beta-access/email.ts              FOUNDER-OWNED
    src/modules/access/beta-access/email.ts   not protected

That is the D9 hazard again. D9's fix does not cover it — D9 stated the rule on
the file TYPE (`src/modules/*.tsx`) and email.ts is a `.ts`.

BUT THIS TIME THE NET HELD, AND THE DIFFERENCE IS WORTH KEEPING

D9's globs NARROWED silently ("matches less than it used to" is undetectable).
These three are patterns that match a named file or a specific directory, so the
move left them matching NOTHING — and CTL-035 failed with all three, by name and
line number, before any of this was fixed. Detectable beats silent, and that is
an argument for naming paths precisely rather than globbing broadly.

All three are re-pointed in THIS commit, so no window exists in which the signup
gate is disarmed. Verified after: CTL-035 back to 0 dead / 155 live, and
is_protected answers FOUNDER-OWNED at the new address.

The trailer above is the standing KAN-415 extraction authorisation, and it is
load-bearing precisely because it is dangerous: a trailer on a commit that moved
a founder-owned file OUT of the protected set would have carried the move
through the gate it was disabling. It is honest here only because email.ts is
byte-identical and is still founder-owned at its new path.

A DEFECT IN THE DoD GATE'S OWN ARCHIVE LIST, FOUND BY DOING THIS

The reference sweep wanted four more files rewritten:

    KAN-419-path-coupling.md · KAN-421-profiles-god-table.md
    KAN-416-module-manifest.md · PLAN-REVALIDATION-2026-07-28.md

Every one says "Spike · research artefact · read-only" in its header, with a
Produced/Run date and the exact commit it was measured at. They are records of a
moment. Rewriting a path inside one falsifies the measurement — which is the
stated reason ARCHIVE_FILES exists.

ARCHIVE_FILES already holds their .json DATA siblings (kan421-profiles-inventory,
kan432-revalidation) and missed the WRITE-UPS. So every extraction has been
rewriting the prose while protecting the data, and the two halves of the same
measurement now disagree.

Already done, not hypothetical: KAN-419-path-coupling.md says "Produced:
2026-07-27 · Tracked files at scan time: 791" and today holds 19 references to
`src/modules/` — a directory that did not exist when that scan ran. D1..D9 put
them there.

This commit reverts its own rewrites of those four and adds them to
ARCHIVE_FILES so the bleeding stops. Restoring the text D1..D9 already changed
needs git archaeology and is tracked separately.

modules.json: the sed produced NESTED path declarations
(`src/modules/access/beta-access/` inside `src/modules/access/`). The manifest's
own integrity test caught it — collapsed to the module root, as profile's geo/
was in the previous commit.

VERIFICATION

tsc clean; CTL-035 0 dead; CTL-013 armed at the new path; no stale reference to
any of the three old paths in src/, tests/, scripts/, .github/, modules.json or
env-access-baseline.json.

The two check-env-access sandbox cases fail while UNCOMMITTED, for the reason
documented in the previous commit — the harness clones HEAD for the tree but
copies the working-tree baseline over it. They pass once committed.

Refs: KAN-415, KAN-428, KAN-411, KAN-413/CTL-013, KAN-419
