+++ scripts/check-ui-copy-ownership.sh
@@ -62,7 +62,7 @@ is_protected() {
   [[ $f == src/lib/invite-text.ts ]] && return 0
   [[ $f == src/lib/convene/invites/templates.ts ]] && return 0
   [[ $f == src/lib/convene/invites/sms-templates.ts ]] && return 0
-  [[ $f == src/lib/beta-access/email.ts ]] && return 0
+  [[ $f == src/modules/access/beta-access/email.ts ]] && return 0
   [[ $f == src/app/dashboard/profile/affiliation-fields.ts ]] && return 0
   [[ $f == src/app/dashboard/convene/organise/organise-fields.ts ]] && return 0
   # --- design / styling / brand ---
