+++ src/app/admin/beta-queue/actions.ts
@@ -2,7 +2,7 @@
 
 import { revalidatePath } from 'next/cache';
 import { getCurrentAdmin, getAdminServiceClient, logModerationAction } from '@/lib/admin';
-import { sendBetaApprovedEmail } from '@/lib/beta-access/email';
+import { sendBetaApprovedEmail } from '@/modules/access/beta-access/email';
 
 /**
  * KAN-277 (epic KAN-273): approve a queued user into the beta.
