+++ src/app/dashboard/convene/gatherings/[id]/actions.ts
@@ -20,7 +20,7 @@ import { adapterFor } from '@/lib/convene/calendar';
 import { getConnectionForUser } from '@/lib/convene/oauth-connections';
 import { generateRsvpToken, persistQueuedInvite, setInviteeRsvpToken } from '@/lib/convene/invites/repository';
 import { dispatchQueuedInvites } from '@/lib/convene/invites/dispatch';
-import { getAccountStanding, shouldRefuseIssuance } from '@/lib/account-status';
+import { getAccountStanding, shouldRefuseIssuance } from '@/modules/access/account-status';
 // SEC-132: type-only import. `'use server'` files may export only async
 // functions (gotcha #18) — a type import is erased at build, so it adds no
 // runtime export and cannot trip that rule.
