+++ src/app/dashboard/page.tsx
@@ -5,7 +5,7 @@ import Image from 'next/image';
 import { signOut } from '../(auth)/actions';
 import ShareBeta from './share-beta';
 import DashboardWidgets, { type WidgetContext } from './widgets/dashboard-widgets';
-import { betaInviteLink, publicSignupUrl } from '@/lib/beta-access/invite-link';
+import { betaInviteLink, publicSignupUrl } from '@/modules/access/beta-access/invite-link';
 import { isConveneEnabledForCurrentUser } from '@/lib/convene/flags-user';
 import { resolveWidgets, resolveOnboardingState } from '@/lib/dashboard/resolve-widgets';
 import { dismissedForState, type DashboardWidgetState } from '@/lib/dashboard/dismissal';
