+++ src/app/join/route.ts
@@ -1,6 +1,6 @@
 import { NextResponse, type NextRequest } from 'next/server';
 import { env } from '@/modules/platform/env';
-import { INVITE_COOKIE, INVITE_COOKIE_MAX_AGE } from '@/lib/beta-access/invite-cookie';
+import { INVITE_COOKIE, INVITE_COOKIE_MAX_AGE } from '@/modules/access/beta-access/invite-cookie';
 
 /**
  * KAN-337 — beta-invite deep-link. `/join?code=<INVITE_CODE>` is the shareable
