+++ src/app/oauth/authorize/page.tsx
@@ -12,7 +12,7 @@
 import { redirect } from 'next/navigation';
 import { createClient as createSupabaseServer } from '@/modules/platform/supabase-server';
 import { validateAuthorizeRequest, buildErrorRedirect } from '@/modules/oauth-as/lib/authorize';
-import { getAccountStanding } from '@/lib/account-status';
+import { getAccountStanding } from '@/modules/access/account-status';
 import { getConsent } from '@/modules/oauth-as/lib/consents';
 import { clientTrust, redirectHost } from '@/modules/oauth-as/lib/client-trust';
 import { buildAuthorizePath } from '@/modules/oauth-as/consent-flow';
