+++ src/app/waitlist/actions.ts
@@ -4,8 +4,8 @@ import { redirect } from 'next/navigation';
 import { createServiceRoleClient } from '@/modules/platform/supabase-service';
 import { createClient } from '@/modules/platform/supabase-server';
 import { env } from '@/modules/platform/env';
-import { computeAccessTransition } from '@/lib/access-model';
-import { getAccountStanding, shouldRefuseIssuance } from '@/lib/account-status';
+import { computeAccessTransition } from '@/modules/access/access-model';
+import { getAccountStanding, shouldRefuseIssuance } from '@/modules/access/account-status';
 
 /**
  * KAN-336 — redeem a skip-the-waitlist code from the /waitlist page.
