+++ src/modules/oauth-as/consent-flow.ts
@@ -43,7 +43,7 @@
  */
 import { redirect } from 'next/navigation';
 import { createClient as createSupabaseServer } from '@/modules/platform/supabase-server';
-import { getAccountStanding, shouldRefuseIssuance } from '@/lib/account-status';
+import { getAccountStanding, shouldRefuseIssuance } from '@/modules/access/account-status';
 import { issueAuthCode } from '@/modules/oauth-as/lib/codes';
 import { recordConsent } from '@/modules/oauth-as/lib/consents';
 import { getOauthClient } from '@/modules/oauth-as/lib/clients';
