+++ tests/unit/invite-link.test.ts
@@ -16,11 +16,11 @@ jest.mock('@/modules/platform/env', () => ({
     siteUrl: () => mockSiteUrl,
   },
 }));
-jest.mock('@/lib/beta-access/flow', () => ({
+jest.mock('@/modules/access/beta-access/flow', () => ({
   isProdFamily: () => mockIsProdFamily,
 }));
 
-import { betaInviteLink, buildBetaInviteLink } from '@/lib/beta-access/invite-link';
+import { betaInviteLink, buildBetaInviteLink } from '@/modules/access/beta-access/invite-link';
 
 beforeEach(() => {
   mockInviteCode = '';
